Never increase ofs in dasm_link, part two
When we emit code with RSP/RBP using a dynamic register, we might need
to emit another byte. This can cause problems with undeestimating the
required label size (as it did with OPTREX before). So we must compute
the maximum required size in dasm_put, and only decrement in dasm_link
if possible.